+++ example/src/App.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { useCallback, useMemo, useState } from 'react';
+import { useState } from 'react';
 import {
   AirplayButton,
   AutoFocusGuide,
@@ -36,7 +36,7 @@ import {
   THEOplayerView,
 } from 'react-native-theoplayer';
 import { Platform, StatusBar, StyleSheet, useColorScheme, View } from 'react-native';
-import { SourceMenuButton, SOURCES } from './custom/SourceMenuButton';
+import { SourceMenuButton } from './custom/SourceMenuButton';
 import { BackgroundAudioSubMenu } from './custom/BackgroundAudioSubMenu';
 import { PiPSubMenu } from './custom/PipSubMenu';
 import { MediaCacheDownloadButton } from './custom/MediaCacheDownloadButton';
@@ -44,7 +44,7 @@ import { MediaCacheMenuButton } from './custom/MediaCacheMenuButton';
 import { MediaCachingTaskListSubMenu } from './custom/MediaCachingTaskListSubMenu';
 import { RenderingTargetSubMenu } from './custom/RenderingTargetSubMenu';
 import { AutoPlaySubMenu } from './custom/AutoPlaySubMenu';
-import { SafeAreaProvider, SafeAreaView, Edges } from 'react-native-safe-area-context';
+import { Edges, SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
 import { usePresentationMode } from './hooks/usePresentationMode';
 import {
   EzdrmFairplayContentProtectionIntegrationFactory,
@@ -52,6 +52,9 @@ import {
   KeyOSDrmWidevineContentProtectionIntegrationFactory,
 } from '@theoplayer/react-native-drm';
 import { ExtensionMenuButton } from './custom/ExtensionMenuButton';
+import { usePlaylist } from './hooks/usePlaylist';
+import SOURCES from './assets/sources.json';
+import { Source } from './custom/Source';
 
 // Register Ezdrm Fairplay integration
 ContentProtectionRegistry.registerContentProtectionIntegration('customEzdrm', 'fairplay', new EzdrmFairplayContentProtectionIntegrationFactory());
@@ -79,9 +82,7 @@ const playerConfig: PlayerConfiguration = {
     mediaSessionEnabled: true,
     skipForwardInterval: 30,
     skipBackwardInterval: 10,
-    convertSkipToSeek: true,
-    allowLivePlayPause: true,
-    seekToLiveOnResume: true,
+    allowLivePlayPause: false,
   },
   ads: {
     theoads: true,
@@ -98,13 +99,11 @@ export default function App() {
   const [player, setPlayer] = useState<THEOplayer | undefined>(undefined);
   const presentationMode = usePresentationMode(player);
   const isDarkMode = useColorScheme() === 'dark';
+  const { currentSource } = usePlaylist(player, SOURCES as Source[]);
 
   // In PiP presentation mode on NewArch Android, there is an issue where SafeAreayView does not update the edges in time,
   // so explicitly disable them here.
-  const edges: Edges = useMemo(
-    () => (Platform.OS === 'android' && presentationMode === PresentationMode.pip ? [] : ['left', 'top', 'right', 'bottom']),
-    [presentationMode],
-  );
+  const edges: Edges = Platform.OS === 'android' && presentationMode === PresentationMode.pip ? [] : ['left', 'top', 'right', 'bottom'];
 
   const onTheoAdsEvent = (event: TheoAdsEvent) => {
     console.log(event);
@@ -113,7 +112,7 @@ export default function App() {
     }
   };
 
-  const onPlayerReady = useCallback((player: THEOplayer) => {
+  const onPlayerReady = (player: THEOplayer) => {
     setPlayer(player);
     // optional debug logs
     player.addEventListener(PlayerEventType.SOURCE_CHANGE, console.log);
@@ -135,7 +134,7 @@ export default function App() {
     sdkVersions().then((versions) => console.log(`[theoplayer] ${JSON.stringify(versions, null, 4)}`));
 
     player.autoplay = true;
-    player.source = SOURCES[0].source;
+    player.source = currentSource.source;
 
     player.backgroundAudioConfiguration = {
       enabled: true,
@@ -149,7 +148,7 @@ export default function App() {
     };
 
     console.log('THEOplayer is ready');
-  }, []);
+  };
 
   return (
     /**
@@ -178,7 +177,7 @@ export default function App() {
                         <MediaCachingTaskListSubMenu />
                       </MediaCacheMenuButton>
                       {/*This is a custom menu for source selection.*/}
-                      <SourceMenuButton />
+                      <SourceMenuButton sources={SOURCES as Source[]} includeWithLicense={playerConfig.license !== undefined} />
                       {!Platform.isTV && (
                         <>
                           <AirplayButton />
@@ -243,7 +242,7 @@ export default function App() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    // on iOS we cannot stretch an inline playerView to cover the whole screen, otherwise it assumes fullscreen presentationMode.
+    // on iOS, we cannot stretch an inline playerView to cover the whole screen, otherwise it assumes fullscreen presentationMode.
     marginHorizontal: Platform.select({ ios: 2, default: 0 }),
     alignItems: 'center',
     justifyContent: 'center',