+++ ios/THEOplayerRCTBridge.m
@@ -315,3 +315,12 @@ RCT_EXTERN_METHOD(setAdTagParameters:(nonnull NSNumber *)node
 
 @end
 
+// ----------------------------------------------------------------------------
+// MediaControl Module
+// ----------------------------------------------------------------------------
+@interface RCT_EXTERN_REMAP_MODULE(THEORCTMediaControlModule, THEOplayerRCTMediaControlAPI, RCTEventEmitter)
+
+RCT_EXTERN_METHOD(setHandler:(nonnull NSNumber *)node
+                  action:(nonnull NSString *)action)
+
+@end