+++ src/internal/adapter/THEOplayerWebAdapter.ts
@@ -84,6 +84,10 @@ export class THEOplayerWebAdapter extends DefaultEventDispatcher<PlayerEventMap>
     }
   }
 
+  get mediaControl() {
+    return this._mediaSession?.mediaControlAdapter;
+  }
+
   get abr(): ABRConfiguration | undefined {
     return this._player?.abr as ABRConfiguration | undefined;
   }